requirements.txt
azure-storage-blob
flask
markdown2
openai
psycopg2-binary
python-dotenv ~= 0.19
requests ~= 2.27
whitenoise